+++ src/lib/util/handleTripFinish.ts
@@ -12,9 +12,8 @@ import { handlePassiveImplings } from '../implings';
 import { triggerRandomEvent } from '../randomEvents';
 import { getUsersCurrentSlayerInfo } from '../slayer/slayerUtil';
 import { ActivityTaskOptions } from '../types/minions';
-import { channelIsSendable, makeComponents } from '../util';
+import { buildClueButtons, channelIsSendable, makeComponents } from '../util';
 import {
-	buildClueButtons,
 	makeAutoContractButton,
 	makeBirdHouseTripButton,
 	makeNewSlayerTaskButton,
