+++ src/mahoji/lib/abstracted_commands/openCommand.ts
@@ -6,10 +6,9 @@ import { Bank, LootTable } from 'oldschooljs';
 import { PerkTier } from '../../../lib/constants';
 import { allOpenables, UnifiedOpenable } from '../../../lib/openables';
 import { ItemBank } from '../../../lib/types';
-import { makeComponents } from '../../../lib/util';
+import { buildClueButtons, makeComponents } from '../../../lib/util';
 import { stringMatches } from '../../../lib/util/cleanString';
 import getOSItem, { getItem } from '../../../lib/util/getOSItem';
-import { buildClueButtons } from '../../../lib/util/globalInteractions';
 import { handleMahojiConfirmation } from '../../../lib/util/handleMahojiConfirmation';
 import { makeBankImage } from '../../../lib/util/makeBankImage';
 import { patronMsg, updateGPTrackSetting } from '../../mahojiSettings';
